Remove state attribute from MCTSAgent, fix search
Refactored MCTSAgent and simulator to have the game state as a
parameter to search.  Previously it had a reference to the actual
game state and was responsible for making copies, which was
an unnecessary amount of access.
Now simulator creates a state copy and passes it to search, which
the agent is free to manipulate.
An unintended side-effect is that the game loop appears to be
running without crashing now.  Unsure if everything is correct,
but at least the game runs and places pieces.