Move BOARD_SIZE to a better spot for clarity
This changes a number of references to it, including the static
generate_winning_positions method in TTTBoard.  Removed the .__func__
workaround and put the call at the bottom of the module.
This seemed a cleaner place to put it than a conditional that checks
for the class attribute in the initializer.

Add some styling to the view and pseudo code the button logic in the
controller.  This logic requires a check in NestedTTT.take_action if
an attempted action is legal, which was added as well.  Simulator
should use this information as well, to prevent cheating or
malfunctioning agents as well as players.