Add ArgumentParser and update the README
Run python main.py -h to see available options.  Currently the
parser hardcodes expecting two agents, but this can probably be
replaced in the future using a single argment and nargs = '*'.
That would require some additional checking to ensure the number
of agents matches the expected number of agents for a game.