Rewrote docstrings, enforce ABCs, extend Simulator
Rewrote or edited almost all of the project's docstrings to be
clearer and provide more context when reviewing the code.

Added inheritance from the ABC class and abstractmethod decorators
to the methods that are expected to be implemented.
This has the advantages of:
- being clearer at a glance that these classes are abstract and which
methods need to be implemented
- and raising exceptions sooner, on the instantiation attempt rather
than when the method is called and raises the NotImplementedError

Simulator has now become another abstract base class.

Its display_game method is now abstract and subclasses override this
depending on their method of display.  ConsoleSimulator now contains
the previous implementation that relies on a __str__ implementation
in the game, and NoDisplaySimulator does not show anything for
maximum simulation speed.  This should make it easier in the future
to add GUI's to games.

The final change to Simulator is that it now supports multiple game
playthroughs.  The number can be set by command-line arguments.