Move display code to a simulator component
Previously, the Simulator class was subclassed and these classes
only updated the display code.  Displaying the game is not
connected to actually running the game simulation, so I removed the
abstract base class from Simulator and now pass it a display
controller that is calls for these operations.

This keeps the two pieces from being so tightly coupled and prevents
the display classes from inheriting a large amount of
unnecessary attributes and methods.