Move responsibility for action verifying to Game
Previously, Simulator was responsible for checking if an action was
legal and taking it or not.
A side benefit of moving this check to game is that the
progress_game decorator is no longer needed.  Simulator calls Game's
concrete method take_action_if_legal that can handle incrementing the
game turn.