Add in-progress HavannahBoard printing work
The top section now prints properly, the top for where there were
no values but just the topmost hex __ was also invoking the s_west
step, which then threw off the rest of the operations.

The top row of the middle portion also prints properly, but then
on the next portion something goes awry.