Change num_agents to be set by Game subclass
Rather than pass the number of agents provided to some Game,
each subclass is now expected to set its own expected number of
players.  This value is then used to set the num_agents attribute.

This puts the value in a more logical place, and allows for error
checking in the simulator to make sure the number of agents
provided now matches the number of players in a game.

Game and Simulator now provide some error checking based on this new
attribute as well as ACTION.  Hopefully these error messages make the
specific problem clearer if a new game is written incorrectly.