Add GUI display for Havannah
Just as with NestedTTT, a Tkiner display is now the default option
for Havannah.

The interface for DisplayController was also changed slightly to
pass the game as a parameter.  This was necessary to avoid a
workaround such as recalculating all of the board hex coordinates
when drawing the board the first time.