Update docstrings and new test for updated code
The new test is a simple one to ensure the colors stay in sync with
the state of the game as long as the get_legal_actions method is used.

This new change improves the number of playouts that MCTSAgent can
run for Havannah in .5 seconds by around 200%.  Updating an attribute
of an object is far less costly than constantly creating a new version
of that object, in this case HavannahAction objects.