Use union-find structure to improve win checking!
HexNodes have been updated with additional attributes to convert grid
from a standard graph structure to a union-find structure.

Similar to how neighbors were pre-calculated before the game start,
we now check for corners and edges at HexNode creation time.  When
unions are performed to combine subsets we combine this information
in the representative nodes.  This allows for constant time checking
if a bridge or fork win has occurred.

This and the previous optimizations for get_legal_actions,
pre-calculating neighbors, and deepcopy, have greatly improved the
simulation time for MCTSAgent when playing Havannah.  This
improvement allows the agent to run enough playouts in 4-5 seconds of
allotted time to break the threshold of the action space size of
Havannah enough to make directed moves towards winning the game.