Change the GUI display's to be internal to games
Similar to how the number of players was internalized, the
game-specific display is now stored in the game instead of being
imported and passed to the simulator by main.  The same end result,
but there is less need for main to change if there is a change made
to the game or its display.