Add logging capabilites to project
The main module now configures the root logger for use by the rest of
the modules.  This is now the expected way to display both
informational and debug messages to the user, instead of calling print.

Simulator shows the expected way to utilize this setup, by calling
getLogger(__name__) to get a module specific logger.

Call the logger's info method for messages that the user may want to
see during normal execution.  The debug method should be used for
messages that will only be written to the log file and provide more
in-depth information.