Add GUI display for MCTSAgent
A number of the interfaces have adjusted slightly to handle this
change, such as the simulator now accepting an argument to let it
know if the game is using a GUI display so that it knows if it
should try to create Agent displays.

GUIDisplayController is now being used for agent displays as well,
and is being passed an agent or a game depending on which is in use.

MCTSDisplay has some in-progress code for drawing node trees commented
out because I have not yet decided how to render them in a clean and
useful way.