Improve how HumanAgent input works
Instead of relying on Action subclasses to do all of the work, now
each action subclass just provides the format prompt string and a
parsing function for their action class.  HumanAgent puts these
together to prompt a user for actions.

This is in preparation for adding a display to HumanAgent that allows
input from the GUI instead of the command-line.