Fix a few minor bugs in MCTSAgent
The exploration portion of UCT's value estimate was the same for all
nodes.  This explains the tendency of the agent to focus its
simulations on a game winning action when it finds one.

Also refactored a few methods to be more simple and clear by removing
some unnecessary return values and branching logic.