Add Gridworld to the project
This is the first step in adding MDP simulation and agents.
Gridworld handles the Game interface and can be passed different Grid
instances, which are currently found in the gridworld_examples.py
file.  This combination of classes will hopefully make it easy to
test gridworld with different levels of uncertainty in the transition
function or different rewards on the same grid.