Move agent display code into Agent base class
Rather than have the simulator store and manage the agent displays,
this is now taken care of by the agent itself.  Most of the logic is
handled by the Agent base class, with a few minor changes to abstract
methods and initializers of the sub classes.

Also, the agents are now capable of resetting back to an initial
state.  The result of this is that initializing a new game playthrough
only requires calling reset on the agent instead of creating a whole
new instance.

One side effect of this change is that agent displays are currently
initialized before the game's display.  This causes an additional Tk
window to appear.  Once Game has undergone the same changes as Agent,
this will no longer occur.