Move display code from Simulator to Game
Following nearly the same pattern as for Agent, the Game base class
and its subclasses now handle their own displays.

This, along with the change to allow games to reset instead of
creating new instances for each run of the game, have cleaned up and
simplified most of the methods in Simulator.