Finish MDP and convert Gridworld to use it
Gridworld now implements all of the MDP functionality without the
extra attributes and method from Game that were not necessary.

Much of the functionality is similar, especially the display code, but
some of the functions have different return values and deserved
different names.

The gridworld examples now have functions that generate the different
gridworld examples based on the desired display type.  Without this
change, each of them was popping up windows at time of import.