Update parser and setup logic for MDPs
The parser now uses two different subparsers, one each for games and
MDPs.  The main function now has diverging setup logic based on which
subparser was invoked.

Simulator was also changed to make all of its methods static.  Since
it was no longer relying on internal state to hold the objects, there
was no need to instantiate the class.